requirements.txt
# In this file list all Python packages that your project needs.
# When someone else wants to get started with your project sources, they can
# simple run 'pip install -r requirements.txt' to automatically install them on
# their machine
#
# Every dependency must be on its own line. Note that the '#'-character is used
# as comment character.
#
# python -m pip install -r requirements.txt

fastapi
uvicorn[standard]
nicegui
httpx
python-multipart
python-dotenv
pymssql
Pillow
pillow-heif
google-genai
geopy
